// uptimer-client/src/interfaces/monitor.interface.ts
export interface IPagination {
  start: number;
  end: number;
}